+++ spec/dummy/client/entries/app-kitten.js
@@ -23,7 +23,7 @@ import DropdownButton from 'kitten/components/dropdowns/dropdown-button'
 // Form
 import { Checkbox } from 'kitten/components/form/checkbox'
 import { FormAmountAndCurrency } from 'kitten/components/form/form-amount-and-currency'
-import FormPhoneNumber from 'kitten/components/form/form-phone-number'
+import { FormPhoneNumber } from 'kitten/components/form/form-phone-number'
 import { Label } from 'kitten/components/form/label'
 import { RadioButton } from 'kitten/components/form/radio-button'
 import Search from 'kitten/components/form/search'